requirements.txt
streamlit
langchain
faiss-cpu
tqdm
python-dotenv
PyPDF2
openai
google-search-results
pypdf
docx2txt
langchain-community
langchain-huggingface
langchain-core
langchain-groq
serpapi